pyproject.toml
[tool.poetry]
name = "sw-api"
version = "0.1.0"
description = ""
authors = ["Luci_shch"]
readme = "README.md"
packages = [{include = "sw_api"}]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.8.4"
aiosignal = "^1.3.1"
asyncpg = "^0.27.0"
attrs = "^23.1.0"
charset-normalizer = "^3.1.0"
frozenlist = "^1.3.3"
greenlet = "^2.0.2"
idna = "^3.4"
more-itertools = "^9.1.0"
multidict = "^6.0.4"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.15"
typing-extensions = "^4.6.2"
yarl = "^1.9.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
